refactor: utility
