+++ packages/ghmattimysql/src/server/types/cfxCallback.ts
@@ -1,0 +1,5 @@
+interface CFXCallback {
+  (args: any): void;
+};
+
+export default CFXCallback;
