+++ packages/ghmattimysql/src/server/utility/safeInvoke.ts
@@ -1,0 +1,11 @@
+import CFXCallback from "../types/cfxCallback";
+
+function safeInvoke(callback: CFXCallback, args: any) {
+  if (typeof callback === 'function') {
+    setImmediate(() => {
+      callback(args);
+    });
+  }
+}
+
+export default safeInvoke;
