+++ packages/ghmattimysql/src/server/utility/sanitizeTransactionInput.ts
@@ -1,0 +1,21 @@
+import { TransactionQueries, TransactionQuery } from "../mysql/transactionQueries";
+import convertTransactionLegacyQueries from "./convertTransactionLegacyQueries";
+import QueryStorage from '../queryStorage';
+
+function sanitizeTransactionInput(querys, params, callback: any): [TransactionQueries, any] {
+  let sqls: TransactionQueries = [];
+  let cb = callback;
+  // start by type-checking and sorting the data
+  const values = (typeof params === 'function') ? [] : params;
+  sqls = querys.map((query: string | number | TransactionQuery) => {
+    if (typeof query === 'string' || typeof query === 'number') { 
+      return { query: QueryStorage.get(query), values };
+    } // we got a Type: TransactionQuery, should actually check that, but we don't
+    return { query: QueryStorage.get(query.query), values: query.values };
+  });
+  if (typeof params === 'function') cb = params;
+  sqls = convertTransactionLegacyQueries(sqls);
+  return [sqls, cb];
+}
+
+export default sanitizeTransactionInput;
