+++ packages/ghmattimysql/src/server/utility/typeCast.ts
@@ -1,0 +1,35 @@
+import { BINARY } from 'mysql/lib/protocol/constants/charsets';
+
+function typeCast(field, next) {
+  let dateString = '';
+  switch (field.type) {
+    case 'DATETIME':
+    case 'DATETIME2':
+    case 'TIMESTAMP':
+    case 'TIMESTAMP2':
+    case 'NEWDATE':
+    case 'DATE':
+      dateString = field.string();
+      if (field.type === 'DATE') dateString += ' 00:00:00';
+      return (new Date(dateString)).getTime();
+    case 'TINY':
+      if (field.length === 1) {
+        return (field.string() !== '0');
+      }
+      return next();
+    case 'BIT':
+      return Number(field.buffer()[0]);
+    case 'TINY_BLOB':
+    case 'MEDIUM_BLOB':
+    case 'LONG_BLOB':
+    case 'BLOB':
+      if (field.packet.charsetNr === BINARY) {
+        return [...field.buffer()];
+      } 
+      return field.string();
+    default:
+      return next();
+  }
+}
+
+export default typeCast;
